+++ shuup/front/middleware.py
@@ -27,12 +27,18 @@ from shuup.core.utils.users import (
 from shuup.front.basket import get_basket
 from shuup.front.utils.user import is_admin_user
 
+try:
+    from django.utils.deprecation import MiddlewareMixin
+except ImportError:
+    MiddlewareMixin = object
+
+
 __all__ = ["ProblemMiddleware", "ShuupFrontMiddleware"]
 
 ProblemMiddleware = ExceptionMiddleware  # This class is only an alias for ExceptionMiddleware.
 
 
-class ShuupFrontMiddleware(object):
+class ShuupFrontMiddleware(MiddlewareMixin):
     """
     Handle Shuup specific tasks for each request and response.